.github/scripts/requirements.txt
easyargs == 0.9.4
PyYAML == 5.4.1
